moved animation code to own module
